Add to the logic for groups and authorized user and also add a new route to the sidebar

+++ backend/controller/authorizedUserCtrl.js
@@ -6,35 +6,35 @@ const AuthorizedUser = require("../models/AuthorizedUser");
 const authorizedUserController = {
 
     // add
-    addAuthorizedUser: asyncHandler(async(req,res)=> {
-        const {expiresAt, encryptedPassword, iv, authorizedId} = req.body
-         if(!iv || !encryptedPassword){
+    addAuthorizedUser: asyncHandler(async (req, res) => {
+        const { expiresAt, encryptedPassword, iv, authorizedId } = req.body
+        if (!iv || !encryptedPassword) {
             return res.status(400).json({
                 message: "Please ensure you encrypt the password"
             })
-         }
-         if(!authorizedId){
+        }
+        if (!authorizedId) {
             return res.status(400).json({
                 message: "Ensure that you select a user to authorize"
             })
-         }
+        }
 
-         const authorizedUser =  await AuthorizedUser.create({
+        const authorizedUser = await AuthorizedUser.create({
             authorizedId,
             expiresAt,
             encryptedPassword,
             iv,
             ownerId: req.user.id,
 
-         })
+        })
 
-         res.status(201).json({
+        res.status(201).json({
             message: "Authorized User successfully"
-         })
+        })
     }),
-    
+
     // get
-    getAuthorizedUsers: asyncHandler(async(req,res)=> {
+    getAuthorizedUsers: asyncHandler(async (req, res) => {
         const authorizedUsers = await AuthorizedUser.find({
             ownerId: req.user.id
         }).lean();
@@ -45,30 +45,49 @@ const authorizedUserController = {
 
 
     // delete
-    deleteAuthorizedUser: asyncHandler(async(req,res)=> {
-        const {authorizedId} = req.params
-        const deletedAuthorizedUser = await AuthorizedUser.findByIdAndDelete({ownerId: req.user.id, _id:authorizedId }).lean();
+    deleteAuthorizedUser: asyncHandler(async (req, res) => {
+        const { authorizedId } = req.params
+        if (!mongoose.Types.ObjectId.isValid(authorizedId)) {
+            res.status(400);
+            throw new Error("Invalid group ID");
+        }
+        const deletedAuthorizedUser = await AuthorizedUser.findByIdAndDelete({ ownerId: req.user.id, _id: authorizedId }).lean();
+
+        if (!deletedAuthorizedUser) {
+            res.status(404);
+            throw new Error("Group not found or unauthorized");
+        }
+
 
         res.status(200).json({
             message: "Authorized user removed successfully",
             deletedAuthorizedUser
         })
     }),
-    
+
     // edit
-    editAuthorizedUser: asyncHandler(async(req,res)=> {
-         const {authorizedId} = req.params
-        const authorizedUser = await AuthorizedUser.findOne({ownerId: req.user.id, _id:authorizedId });
+    editAuthorizedUser: asyncHandler(async (req, res) => {
+        const { authorizedId } = req.params
+        if (!mongoose.Types.ObjectId.isValid(authorizedId)) {
+            res.status(400);
+            throw new Error("Invalid group ID");
+        }
+
+        const authorizedUser = await AuthorizedUser.findOne({ ownerId: req.user.id, _id: authorizedId });
 
         res.status(200).json({
             message: "Authorized user info updated successfully"
         })
     }),
-    toggleAuthorizedUser: asyncHandler(async(req,res)=> {
-        const {authorizedId} = req.params
-        const authorizedUser = await AuthorizedUser.findOne({ownerId: req.user.id, _id:authorizedId });
+    toggleAuthorizedUser: asyncHandler(async (req, res) => {
+        const { authorizedId } = req.params
+        if (!mongoose.Types.ObjectId.isValid(authorizedId)) {
+            res.status(400);
+            throw new Error("Invalid group ID");
+        }
+        const authorizedUser = await AuthorizedUser.findOne({ ownerId: req.user.id, _id: authorizedId });
 
-        if(authorizedUser){
+        if (authorizedUser) {
             authorizedUser.authorized = !authorizedUser.authorized;
         }
